requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0

# API Clients
openai>=1.3.0

# Data Processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0
python-dotenv>=1.0.0

# Language Processing
langdetect>=1.0.9
nltk>=3.8.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.9.0
flake8>=6.1.0